+++ lib/data/datasources/remote/user_datasource.dart
@@ -42,7 +42,7 @@ class UserDatasource {
       String? username,
       // String? password,
       bool? geolocEnabled,
-      int? prefGeolocRadius,
+      double? prefGeolocRadius,
       double? lat,
       double? lng}) async {
     Map<String, dynamic> data = {
